% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awtew.R
\docType{data}
\name{awtew}
\alias{awtew}
\title{The text from All's Well That Ends Well}
\format{A charcter vector with 24235 items}
\usage{
data(awtew)
}
\description{
A charcter vector for testing and benching marking from
\url{http://shakespeare.mit.edu/allswell/full.html}
}
\keyword{datasets}
