% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hs_platform}
\alias{hs_platform}
\title{Utility function to determine what CPU optimizations can be used}
\usage{
hs_platform()
}
\value{
list with cpu features and platform.
}
\description{
Utility function to determine what CPU optimizations can be used
}
\examples{
hs_platform()
}
