% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hs_str_count}
\alias{hs_str_count}
\title{Count the number of matches in a string.}
\usage{
hs_str_count(string, pattern)
}
\arguments{
\item{string}{Input character vector.}

\item{pattern}{Pattern to look for.}
}
\value{
An integer vector.
}
\description{
Count the number of matches in a string.
}
\examples{
fruit <- c("apple", "banana", "pear", "pineapple")
hs_str_count(fruit, "a")
hs_str_count(fruit, "p")
hs_str_count(fruit, "e")
}
