% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hs_str_detect}
\alias{hs_str_detect}
\title{Detect if the pattern is in a string}
\usage{
hs_str_detect(Rstring, Rpattern)
}
\arguments{
\item{string}{Input character vector.}

\item{pattern}{Pattern to look for.}
}
\value{
A logical vector.
}
\description{
Detect if the pattern is in a string
}
\examples{
fruit <- c("apple", "banana", "pear", "pineapple")
hs_str_detect(fruit, "app")
hs_str_detect(fruit, "p")
}
