% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hs_str_locate}
\alias{hs_str_locate}
\title{Locate the first instance of a pattern}
\usage{
hs_str_locate(Rstring, Rpattern)
}
\arguments{
\item{string}{Input character vector.}

\item{pattern}{Pattern to look for.}
}
\value{
A integer matrix.
}
\description{
Locate the first instance of a pattern
}
\examples{
fruit <- c("apple", "banana", "pear", "pineapple")
hs_str_locate(fruit, "app")
hs_str_locate(fruit, "p")
}
