% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hs_str_locate_all}
\alias{hs_str_locate_all}
\title{Locate the all instances of a pattern}
\usage{
hs_str_locate_all(string, pattern)
}
\arguments{
\item{string}{Input character vector.}

\item{pattern}{Pattern to look for.}
}
\value{
A list of integer matrix.
}
\description{
Locate the all instances of a pattern
}
\examples{
fruit <- c("apple", "banana", "pear", "pineapple")
hs_str_locate_all(fruit, "app")
hs_str_locate_all(fruit, "p")
}
