% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hs_str_replace}
\alias{hs_str_replace}
\title{Replace the first instance of a pattern}
\usage{
hs_str_replace(string, pattern, replacement)
}
\arguments{
\item{string}{Input character vector.}

\item{pattern}{Pattern to look for.}

\item{replacement}{String to replace pattern with.}
}
\value{
A charcter vector.
}
\description{
Replace the first instance of a pattern
}
