% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hs_str_replace_all}
\alias{hs_str_replace_all}
\title{Replace each instance of a pattern}
\usage{
hs_str_replace_all(string, pattern, replacement)
}
\arguments{
\item{string}{Input character vector.}

\item{pattern}{Pattern to look for.}

\item{replacement}{String to replace pattern with.}
}
\value{
A charcter vector.
}
\description{
Replace each instance of a pattern
}
\examples{
fruit <- c("apple", "banana", "pear", "pineapple")
hs_str_replace_all(fruit, "app","simi")
hs_str_replace_all(fruit, "p", "P")
}
