% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{hs_str_trim}
\alias{hs_str_trim}
\title{Trim whitespace}
\usage{
hs_str_trim(string, side = c("both", "left", "right"))
}
\arguments{
\item{string}{Input character vector}

\item{side}{string, either "both", "left" or "right"}
}
\value{
character vector
}
\description{
Trim whitespace
}
\examples{
trim <- c(" left", "right  ", "  both  ")
hs_str_trim(trim, "left")
hs_str_trim(trim, "right")
hs_str_trim(trim, "both")
}
