% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hs_str_which}
\alias{hs_str_which}
\title{Detect which strings match a pattern}
\usage{
hs_str_which(string, pattern)
}
\arguments{
\item{string}{Input character vector.}

\item{pattern}{Pattern to look for.}
}
\value{
A integer vector of length of matches.
}
\description{
Detect which strings match a pattern
}
\examples{
fruit <- c("apple", "banana", "pear", "pineapple")
hs_str_which(fruit, "a")
hs_str_which(fruit, "p")
}
