% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hs_verion}
\alias{hs_verion}
\title{Utility function to determine what version of hyperscan library}
\usage{
hs_verion()
}
\value{
string with library version.
}
\description{
Utility function to determine what version of hyperscan library
}
\examples{
hs_verion()
}
